Added no-token api route and fixed baseURL
the /no-token route from account api applies to the login page only. If there's no token, the user can proceed; otherwise, they are redirected back to the homepage.

also fixed the issue with axios.default.baseURL. This allows for components to not include the entire route when calling to the backend server.